Initialize Assets
Initializing assets based on the designer's documents. Had some troubles since she wrote down some attributes that we have not implemented.

TODO
- Need to introduce the need name change to all need systems
- Resolve bugs with map designer
- Implement attributes from designer

+++ Space-Zoologist/Assets/Scripts/UI/Store/Item.cs
@@ -9,14 +9,14 @@ public class Item : ScriptableObject
     public string ID => id;
     public ItemType Type => type;
     public string ItemName => itemName;
-    public int Price => price;
+    public float Price => price;
     public string Description => description;
     public Sprite Icon => icon;
 
     [SerializeField] private string id = default;
     [SerializeField] private ItemType type = default;
     [SerializeField] private string itemName = default;
-    [SerializeField] private int price = default;
+    [SerializeField] private float price = default;
     [SerializeField] private string description = default;
     [SerializeField] private Sprite icon = default;
